Add files via upload
The 21.2 challenge is to use machine learning to predict customers ready for a device upgrade. TracFone offers customers mobile phones to go with the plans from feature phones, to Apple iPhones and Android smartphones. An aging device where the battery is loosing its ability to retain charge may keep running out of battery when the customer needs it most. Meanwhile certain customer’s usage and lifestyles may be better catered with the latest mobile phone. Furthermore a grandmother who is accustomed to her mobile phone may not be comfortable getting a new device.
If we can accurately predict when a customer is ready for an upgrade, we can ensure that only the customers looking to upgrade their device do in fact get the offer to upgrade their device. Proactively anticipating and serving our customer’s needs for a new device will ensure a good experience for our customers.